Separate complicated header parts on search results
