+++ frontend/src/components/entry/SearchResults.tsx
@@ -1,13 +1,7 @@
-import {
-  AdvancedSearchResultAttrInfoFilterKeyEnum,
-  AdvancedSearchResultValue,
-} from "@dmm-com/airone-apiclient-typescript-fetch";
-import FilterAltIcon from "@mui/icons-material/FilterAlt";
-import FilterListIcon from "@mui/icons-material/FilterList";
+import { AdvancedSearchResultValue } from "@dmm-com/airone-apiclient-typescript-fetch";
 import {
   Box,
   Checkbox,
-  IconButton,
   List,
   ListItem,
   Pagination,
@@ -17,46 +11,24 @@ import {
   TableBody,
   TableCell,
   TableContainer,
-  TableHead,
   TableRow,
   Typography,
 } from "@mui/material";
 import { styled } from "@mui/material/styles";
-import React, { ChangeEvent, FC, useState, useReducer } from "react";
-import { useHistory, useLocation, Link } from "react-router-dom";
+import React, { FC, useMemo } from "react";
+import { Link } from "react-router-dom";
 
-import { SearchResultControlMenuForEntry } from "./SearchResultControlMenuForEntry";
-import { SearchResultControlMenuForReferral } from "./SearchResultControlMenuForReferral";
+import { SearchResultsTableHead } from "./SearchResultsTableHead";
 
 import { entryDetailsPath } from "Routes";
 import { AttributeValue } from "components/entry/AttributeValue";
-import { SearchResultControlMenu } from "components/entry/SearchResultControlMenu";
-import {
-  AttrFilter,
-  AttrsFilter,
-  formatAdvancedSearchParams,
-} from "services/entry/AdvancedSearch";
+import { AttrsFilter } from "services/entry/AdvancedSearch";
 
 const StyledBox = styled(Box)({
   display: "table",
   overflowX: "inherit",
   padding: "24px",
 });
-
-const HeaderBox = styled(Box)({
-  display: "flex",
-  justifyContent: "space-between",
-  alignItems: "center",
-});
-
-const StyledIconButton = styled(IconButton)(({ theme }) => ({
-  color: theme.palette.primary.contrastText,
-}));
-
-const StyledTableCell = styled(TableCell)(({ theme }) => ({
-  color: theme.palette.primary.contrastText,
-  minWidth: "300px",
-}));
 
 const StyledTableRow = styled(TableRow)({
   "&:nth-of-type(odd)": {
@@ -98,193 +70,21 @@ export const SearchResults: FC<Props> = ({
   bulkOperationEntryIds,
   handleChangeBulkOperationEntryId,
 }) => {
-  const location = useLocation();
-  const history = useHistory();
-
-  const isFiltered: Record<string, boolean> = Object.fromEntries(
-    Object.keys(defaultAttrsFilter ?? {}).map((attrName: string) => {
-      const attrFilter = defaultAttrsFilter[attrName];
-      switch (attrFilter.filterKey) {
-        case AdvancedSearchResultAttrInfoFilterKeyEnum.EMPTY:
-        case AdvancedSearchResultAttrInfoFilterKeyEnum.NON_EMPTY:
-        case AdvancedSearchResultAttrInfoFilterKeyEnum.DUPLICATED:
-          return [attrName, true];
-        case AdvancedSearchResultAttrInfoFilterKeyEnum.TEXT_CONTAINED:
-          return [attrName, attrFilter.keyword !== ""];
-        case AdvancedSearchResultAttrInfoFilterKeyEnum.TEXT_NOT_CONTAINED:
-          return [attrName, attrFilter.keyword !== ""];
-      }
-      return [attrName, false];
-    })
-  );
-
-  const [entryFilter, entryFilterDispatcher] = useReducer(
-    (
-      _state: string,
-      event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-    ) => event.target.value,
-    defaultEntryFilter ?? ""
-  );
-  const [referralFilter, referralFilterDispatcher] = useReducer(
-    (
-      _state: string,
-      event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-    ) => event.target.value,
-    defaultReferralFilter ?? ""
-  );
-  const [attrsFilter, setAttrsFilter] = useState<AttrsFilter>(
-    defaultAttrsFilter ?? {}
-  );
-
-  const [attributeMenuEls, setAttributeMenuEls] = useState<{
-    [key: string]: HTMLButtonElement | null;
-  }>({});
-  const [entryMenuEls, setEntryMenuEls] = useState<HTMLButtonElement | null>(
-    null
-  );
-  const [referralMenuEls, setReferralMenuEls] =
-    useState<HTMLButtonElement | null>(null);
-
-  const handleSelectFilterConditions =
-    (attrName?: string) =>
-    (
-      attrFilter?: AttrFilter,
-      overwriteEntryName?: string,
-      overwriteReferral?: string
-    ) => {
-      const _attrsFilter =
-        attrName != null && attrFilter != null
-          ? { ...attrsFilter, [attrName]: attrFilter }
-          : attrsFilter;
-
-      const newParams = formatAdvancedSearchParams({
-        attrsFilter: _attrsFilter,
-        entryName: overwriteEntryName ?? entryFilter,
-        referralName: overwriteReferral ?? referralFilter,
-        baseParams: new URLSearchParams(location.search),
-      });
-
-      // simply reload with the new params
-      history.push({
-        pathname: location.pathname,
-        search: "?" + newParams.toString(),
-      });
-      history.go(0);
-    };
-
-  const handleUpdateAttrFilter =
-    (attrName: string) => (attrFilter: AttrFilter) => {
-      setAttrsFilter((attrsFilter) => ({
-        ...attrsFilter,
-        [attrName]: attrFilter,
-      }));
-    };
+  const attrNames = useMemo(() => {
+    return Object.keys(defaultAttrsFilter);
+  }, [defaultAttrsFilter]);
 
   return (
     <Box display="flex" flexDirection="column">
       <StyledBox>
         <TableContainer component={Paper}>
           <Table id="table_result_list">
-            <TableHead>
-              <TableRow sx={{ backgroundColor: "primary.dark" }}>
-                <TableCell sx={{ witdh: "80px" }} />
-                <StyledTableCell sx={{ outline: "1px solid #FFFFFF" }}>
-                  <HeaderBox>
-                    <Typography>エントリ名</Typography>
-                    <StyledIconButton
-                      onClick={(e) => {
-                        setEntryMenuEls(e.currentTarget);
-                      }}
-                    >
-                      {defaultEntryFilter ? (
-                        <FilterAltIcon />
-                      ) : (
-                        <FilterListIcon />
-                      )}
-                    </StyledIconButton>
-                    <SearchResultControlMenuForEntry
-                      entryFilter={entryFilter}
-                      anchorElem={entryMenuEls}
-                      handleClose={() => setEntryMenuEls(null)}
-                      entryFilterDispatcher={entryFilterDispatcher}
-                      handleSelectFilterConditions={handleSelectFilterConditions()}
-                      handleClear={() =>
-                        handleSelectFilterConditions()(undefined, "")
-                      }
-                    />
-                  </HeaderBox>
-                </StyledTableCell>
-                {Object.keys(attrsFilter).map((attrName) => (
-                  <StyledTableCell key={attrName}>
-                    <HeaderBox>
-                      <Typography>{attrName}</Typography>
-                      <StyledIconButton
-                        onClick={(e) => {
-                          setAttributeMenuEls({
-                            ...attributeMenuEls,
-                            [attrName]: e.currentTarget,
-                          });
-                        }}
-                      >
-                        {isFiltered[attrName] ?? false ? (
-                          <FilterAltIcon />
-                        ) : (
-                          <FilterListIcon />
-                        )}
-                      </StyledIconButton>
-                      <SearchResultControlMenu
-                        attrFilter={attrsFilter[attrName]}
-                        anchorElem={attributeMenuEls[attrName]}
-                        handleClose={() =>
-                          setAttributeMenuEls({
-                            ...attributeMenuEls,
-                            [attrName]: null,
-                          })
-                        }
-                        handleSelectFilterConditions={handleSelectFilterConditions(
-                          attrName
-                        )}
-                        handleUpdateAttrFilter={handleUpdateAttrFilter(
-                          attrName
-                        )}
-                      />
-                    </HeaderBox>
-                  </StyledTableCell>
-                ))}
-                {hasReferral && (
-                  <StyledTableCell sx={{ outline: "1px solid #FFFFFF" }}>
-                    <HeaderBox>
-                      <Typography>参照エントリ</Typography>
-                      <StyledIconButton
-                        onClick={(e) => {
-                          setReferralMenuEls(e.currentTarget);
-                        }}
-                      >
-                        {defaultReferralFilter ? (
-                          <FilterAltIcon />
-                        ) : (
-                          <FilterListIcon />
-                        )}
-                      </StyledIconButton>
-                      <SearchResultControlMenuForReferral
-                        referralFilter={referralFilter}
-                        anchorElem={referralMenuEls}
-                        handleClose={() => setReferralMenuEls(null)}
-                        referralFilterDispatcher={referralFilterDispatcher}
-                        handleSelectFilterConditions={handleSelectFilterConditions()}
-                        handleClear={() =>
-                          handleSelectFilterConditions()(
-                            undefined,
-                            undefined,
-                            ""
-                          )
-                        }
-                      />
-                    </HeaderBox>
-                  </StyledTableCell>
-                )}
-              </TableRow>
-            </TableHead>
+            <SearchResultsTableHead
+              hasReferral={hasReferral}
+              defaultEntryFilter={defaultEntryFilter}
+              defaultReferralFilter={defaultReferralFilter}
+              defaultAttrsFilter={defaultAttrsFilter}
+            />
             <TableBody>
               {results.map((result) => (
                 <StyledTableRow key={result.entry.id}>
@@ -307,7 +107,7 @@ export const SearchResults: FC<Props> = ({
                       {result.entry.name}
                     </Box>
                   </TableCell>
-                  {Object.keys(attrsFilter).map((attrName) => (
+                  {attrNames.map((attrName) => (
                     <TableCell key={attrName}>
                       {(() => {
                         const info = result.attrs[attrName];
